Modified ride.cs to handle the searchRand as a double not as an int. Required for the search rank algorithm.

+++ grabbaride/trunk/GrabbaRide.Database/Ride.cs
@@ -4,7 +4,7 @@ namespace GrabbaRide.Database
 {
     public partial class Ride
     {
-        public int SearchRank { get; set; }
+        public double SearchRank { get; set; }
 
         public TimeSpan DepartureTime
         {
